Add camera handler component
- add new camera component with all of the recording functionalities and driver setup
